requirements.txt
langchain-community>=0.3.1
langgraph>=0.2.32
starlette>=0.39.2
uvicorn[standard]>=0.31.0
twilio>=9.2.3
gunicorn>=23.0.0
python-dotenv>=1.0.1
g711==1.6.5
